added support for floating point numbers as cross values

+++ rasteroperations/crossrasters.h
@@ -29,7 +29,7 @@ private:
 
     bool crossNoRaster(const BoundingBox& box);
     bool crossWithRaster(const  BoundingBox& box);
-    QString determineCrossId(quint32 v1, quint32 v2) const;
+    QString determineCrossId(double v1, double v2) const;
     void checkUndef(double &v1, double &v2);
 };
 }
